+++ src/widget/widget.rs
@@ -142,6 +142,17 @@ pub trait Widget {
 
         self.clear_invalidate();
     }
+
+    /// Sets an alternate context with a drawing offset and size scale, then calls the underlying draw
+    /// routine.  If this is _not_ the desired effect, this routine can be overridden, but this
+    /// will handle the offset drawing for 90% of the `Widget` draw routines.
+    fn draw_with_offset(&mut self, c: Context, g: &mut G2d, clip: &DrawState, point_offset: Point) {
+        self.draw(
+            c.trans(point_offset.x as f64, point_offset.y as f64),
+            g,
+            clip,
+        );
+    }
 }
 
 /// This is the `CanvasWidget`, which contains a top-level widget for display.  It does